### 2024-01-12 #6221 @editorialbot
[PRE REVIEW]: HARE: A Python workflow for analyzing genomic feature enrichment in GWAS datasets
**Submitting author:** @ossmith (<a href="http://orcid.org/0000-0001-5435-2982">Olivia Smith</a>)
**Repository:** https://github.com/ossmith/HARE
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.1.0
