### 2024-01-17 #6221 @Kevin-Mattheus-Moerman
Just copying the notes to the editor here, as there is a prior/related publication (https://doi.org/10.1126/science.adf8009):

> Notes to editor: Version 1.0.0 of this pipeline was used in Kun, et al., Science (2023). Prior to submission to JOSS, we have added extensive functionality, improved the usability, and included unit testing and further documentation. Thus this submission would be the publication of the first production version of the software. An open question of interest in genomics is whether the genetic basis of a set of given traits also happen to lie within regions of the genome associated with functional or evolutionary consequences of interest. This software uses genome-wide association summary statistics to find phenotype-associated genetic features and test for enrichment against these elements of interest (e.g. human accelerated regions). Our software is fast, flexible, and broadly applicable to a rapidly-growing field.
