### 2024-01-17 #6221 @Kevin-Mattheus-Moerman
@ossmith thanks for this submission. I have just triggered a scope query. This means I am seeking the help of the domain experts in the editorial board to see if this work is in scope for JOSS. In particular if the functionality/size of the package is significant enough, also in light of the prior publication. This scope review should take about 2 weeks to complete.
