### 2024-02-05 #6221 @jromanowska
Hi, @nfb1993 - this is the pre-review thread for the paper. Thank you for agreeing to contribute. We will wait for reviewer no.2 to be assigned to start the _review_ thread. In the meantime, you can check [reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
