### 2024-02-14 #6221 @jromanowska
> sorry, I cannot work on JOSS until April and will restart my role as editor then. @abartlett004, would you have time to be a reviewer here?

Thank you for suggesting another person! And sorry - I should've checked first in the JOSS reviewer list. I'm a new editor - still learning :pray:
