### 2024-02-15 #6221 @kdm9
> > @jromanowska sure, I can review this, though it might take me a bit longer than would be ideal
>
> If you are too busy, that's fine. I managed to find two reviewers, so you can relax this time ;) But thank you for the interest. May I keep contact with you, if needed in the future?

Of course. If you end up needing a tiebreaker reviwer please ping me.

And I can give a single overarching review comment for free :smile::  Seems like a rather impressive tool, but could the authors also please confirm that the tool works with non-human data. I see no inherent reason it can't, but the entire documentation assumes human data. Perhaps a quick test with e.g. [AraGWAS](https://aragwas.1001genomes.org) or some other plant dataset would be wise, just to check there are no bugs with non-human data.
