### 2024-02-18 #6359 @nfb1993
Best regards
Nicolas

On Thu, Feb 15, 2024 at 6:34 PM The Open Journals editorial robot <
***@***.***> wrote:

> *Submitting author:* @ossmith <https://github.com/ossmith> (Olivia Smith
> <http://orcid.org/0000-0001-5435-2982>)
> *Repository:* https://github.com/ossmith/HARE
> *Branch with paper.md* (empty if default branch): joss-submission
> *Version:* v1.1.0
> *Editor:* @jromanowska <https://github.com/jromanowska>
> *Reviewers:* @nfb1993 <https://github.com/nfb1993>, @abartlett004
> <https://github.com/abartlett004>
> *Archive:* Pending
> Status
>
> [image: status]
> <https://joss.theoj.org/papers/bdffac35084526179352a807afd9b942>
>
> Status badge code:
>
>
> *Reviewers and authors*:
>
> Please avoid lengthy details of difficulties in the review thread.
> Instead, please create a new issue in the target repository
> <https://github.com/ossmith/HARE> and link to those issues (especially
> acceptance-blockers) by leaving comments in the review thread below. (For
> completists: if the target issue tracker is also on GitHub, linking the
> review thread in the issue or vice versa will create corresponding
> breadcrumb trails in the link target.)
> Reviewer instructions & questions
>
> @nfb1993 <https://github.com/nfb1993> & @abartlett004
> <https://github.com/abartlett004>, your review will be checklist based.
> Each of you will have a separate checklist that you should update when
> carrying out your review.
> First of all you need to run this command in a separate comment to create
> the checklist:
>
> @editorialbot generate my checklist
>
> The reviewer guidelines are available here:
> https://joss.readthedocs.io/en/latest/reviewer_guidelines.html. Any
> questions/concerns please let @jromanowska
> <https://github.com/jromanowska> know.
>
> ✨ *Please start on your review when you are able, and be sure to complete
> your review in the next six weeks, at the very latest* ✨
> Checklists
>
> @nfb1993 <https://github.com/nfb1993>, please create your checklist
> typing: @editorialbot generate my checklist
>
> @abartlett004 <https://github.com/abartlett004>, please create your
> checklist typing: @editorialbot generate my checklist
>
