### 2024-02-28 #6359 @jromanowska
Dear @nfb1993 - how is it going? I can see your generated checklist is formatted wrongly, maybe because you did that via email. Could you edit the comment above and try to fix that or perhaps generate the checklist once more, from GitHub?
