### 2024-03-10 #6359 @nfb1993
Dear editorial team

Thank you for the opportunity to review this paper. Overall, the documentation is well described and the manuscript states clearly the potential of the software. However, there are some boxes that I have not ticked yet, as the repository as it current state lacks of them. Also, I highlight some issues that I encountered when evaluating the software.

Example of usage: Although the commands are clearly stated and the authors provide some testing scripts to verify installation, there is not a friendly example of how to use the software. I highlight that I was able to run a test example made by myself (I used a T2D publicly available GWAS summary statistics).  It would be great if the author could add an small example with some examples of the expected main outputs.

Community guidelines: the author provides where to address potential bugs and issues, however there is not statement for Software contribution i.e. if somebody external can contribute on the code.

Correction necessary on the repository: The necessary dependencies for installation are listed, but the link to version of the human genome cache they recommend ((homo_sapiens_vep_105_GRCh37.tar.gz) ) is wrong.

Correction needed on the software: When trying to run the preranking tool I encountered an error and crashed. The issue was that the time package was not imported. I had to add the line of code “import time” manually to make it work.

Minor correction in article text: In line 81, “femur rations Using HARE, we decided…”, a dot seems missing.

If these points are addressed, I would gladly recommend the software/article for publication.

Side note, not necessary to approve the software: I would not mention Mac Os compatibility, unless it is thoroughly tested. I tried to test the software in an M2 silicon Mac with Ventura OS, I just could not make it run. I was not able to make the conda environment or install ENSEMBL tools. Furthermore, the ENSEMBL tutorial for MAC is outdated and not usable for silicon macs (probably it is for Intel based Macs).  I do not think is the software fault. M2 Macs seem to have problem to run software in perl  (almost not supported there) and certain c++ tools. These problems seem not to be in intel macs, hence, if possible I would try to test the software in a M1-2 Mac to be sure they are compatible.  I was able to run the software very smoothly in a Unix environment.
