### 2024-03-13 #6359 @abartlett004
I mainly agree with Nicolas--the paper and documentation are both well-written and informative, and I would happily recommend the package for publication if the following minor issues are addressed. Each of the checklist boxes I have left unchecked is listed below along with the reason why:

Installation: as Nicolas mentioned, the link to the recommended human genome cache is wrong. I think it is possible that the intended link is something like `https://ftp.ensembl.org/pub/release-105/variation/indexed_vep_cache/homo_sapiens_vep_105_GRCh37.tar.gz`, but please confirm

Community Guidelines: agree with Nicolas as stated above, need guidance for software contribution

Example of Usage: I also recommend the inclusion of a small test dataset of inputs and expected outputs for each of the functions. Perhaps the data from Figure 2 from the paper would be suitable.

Functionality: When a test dataset is included, I will be able to evaluate functionality
