### 2024-04-06 8fe541f
MAJOR - addition of multi-species functionality, example set
- Allows for runs on non-human species.
- Now includes example dataset for users
- Updated VEP to v110, scipy>1.8 (req. Python>3.8)
- Typo corrections in README and paper
