+++ paper/paper.bib
@@ -181,3 +181,50 @@ human-evolved diseases and morphological traits"},
   doi = {10.1093/nar/gkt439},
   url = {https://doi.org/10.1093/nar/gkt439}
 }
+
+@ARTICLE{Zhuang2020,
+  title = {"Weighted Single-Step Genome-Wide Association Study for Growth Traits in Chinese Simmental Beef Cattle"},
+  author = {Zhanwei Zhuang and Lingyang Xu and Jie Yang and Huijiang Gao and Lupei Zhang and Xue Gao and Junya Li and Bo Zhu},
+  journal = {Genes},
+  volume = {11},
+  issue = {2},
+  pages = {189},
+  year = {2020},
+  doi = {10.3390/genes11020189},
+  url = {https://doi.org/10.3390/genes11020189}
+}
+
+@ARTICLE{Atwell2010,
+  title = {"Genome-wide association study of 107 phenotypes in Arabidopsis thaliana inbred lines"},
+  author = {Susanna Atwell and Yu S. Huang and Bjarni J. Vilhjalmsson and Glenda Willems and Matthew Horton and Yan Li and Dazhe Meng and Alexander Platt and Aaron M. Tarone and Tina T. Hu and Rong Jiang and N. Wayan Muliyati and Xu Zhang and Muhammad Ali Amer and Ivan Baxter and Benjamin Brachi and Joanne Chory and Caroline Dean and Marilyne Debieu and Juliette de Meaux and Joseph R. Ecker and Nathalie Faure and Joel M. Kniskern and Jonathan D. G. Jones and Todd Michael and Adnane Nemri and Fabrice Roux and David E. Salt and Chunlao Tang and Marco Todesco and M. Brian Traw and Detlef Weigel and Paul Marjoram and Justin O. Borevitz and Joy Bergelson and Magnus Nordborg},
+  journal = {Nature},
+  volume = {465},
+  pages = {627-631},
+  year = {2010},
+  doi = {10.1038/nature08800},
+  url = {https://doi.org/10.1038/nature08800}
+}
+
+@ARTICLE{Avalos2020,
+  title = {"Genomic regions influencing aggressive behavior in honey bees are defined by colony allele frequencies"},
+  author = {Arian Avalos and Miaoquan Fang and Hailin Pan and Aixa Ramirez Lluch and Alexander E. Lipka and Sihai Dave Zhao and Tugrul Giray and Gene E. Robinson and Guojie Zhang and Matthew E. Hudson},
+  journal = {PNAS},
+  volume = {117},
+  issue = {29},
+  pages = {17135-17141},
+  year = {2020},
+  doi = {10.1073/pnas.1922927117},
+  url = {https://doi.org/10.1073/pnas.1922927117}
+}
+
+@ARTICLE{Sardi2018,
+  title = {"Genome-wide association across Saccharomyces cerevisiae strains reveals substanial variaiton in underlying gene requirements for toxin tolerance"},
+  author = {Maria Sardi and Vaishnavi Paithane and Michael Place and De Elegant Robinson and James Hose and Dana J. Wohlbach and Audrey P. Gasch},
+  journal = {PLOS Genetics},
+  volume = {114},
+  issue = {2},
+  pages = {e1007217},
+  year = {2018},
+  doi = {10.1371/journal.pgen.1007217},
+  url = {https://doi.org/10.1371/journal.pgen.1007217}
+}
