+++ paper/paper.md
@@ -34,7 +34,7 @@ Genomic datasets have grown by orders of magnitude in the last decade and thus r
 # Statement of need
 HARE is a computational pipeline run via command line for analyzing enrichment of element sets, such as the set of human-accelerated regions (HARs), in genomic regions associated with a phenotype of interest. Several genetic enrichment approaches are already available in the literature, such as gene set enrichment analysis (GSEA) [@Subramanian2005; @Mootha2003], stratified linkage disequilibrium score regression (LDSC) [@Finucane2015], and FUMA GWAS [@Wantanabe2017]. However, these approaches are not translatable to studies examining overlaps in custom regions of the genome or when these regions are short, which is known to lead to bias in some approaches [@Finucane2015]. Our pipeline is flexible on either end of the enrichment analysis, both in terms of the regions it utilizes as well as the type of input genomic scan provided, allowing for a completely flexible model for genomic enrichment analysis.
 
-Many tools also require time-consuming data reformatting which HARE minimizes by accepting a variety of summary statistic file formats. In contrast, HARE makes it possible to install the environment with conda and pip, allowing for an immediate run of the complete analysis pipeline. HARE has already been used to analyze over 200 traits in published work as well as ongoing studies presented at conferences [@Kun2023a; @Kun2023b; @Xu2023], exemplifying its usefulness in current and future research.
+Many tools also require time-consuming data reformatting which HARE minimizes by accepting a variety of summary statistic file formats. HARE dependencies can be installed with conda and pip, allowing for an immediate run of the complete analysis pipeline. HARE has already been used to analyze over 200 traits in published work as well as ongoing studies presented at conferences [@Kun2023a; @Kun2023b; @Xu2023], exemplifying its usefulness in current and future research.
 
 # Dependencies
 HARE requires Ensembl's Variant Effect Predictor [@McLaren2016] with the appropriate cache ([VEP Cache Documentation](https://uswest.ensembl.org/info/docs/tools/vep/script/vep_cache.html#cache)) as well as BEDTools [@Quinlan2010]. It also requires the use of `wget` for querying Ensembl BioMart [@Martin2023]. It is recommended that users install dependencies through a combination of conda using the `environment.yml` file included in the repository and pip. Complete instructions for dependency and tool installation can be found in the repository README.
@@ -45,21 +45,23 @@ A visual depiction of the HARE workflow is shown in Figure 1.
 ![HARE visual workflow. **(A)** Summary statistics from a genome-wide association study (GWAS) are filtered to genome-wide significant single nucleotide polymorphisms (SNPs) based on p-value threshold. These are the phenotype-associated (PA) SNPs. **(B)** For each PA SNP, the nearest genetic feature within a given bp distance and its location are annotated. This is called the element set. **(C)** Simulated element sets built from random regions matched on length to the PA features are generated. **(D)** Intersections per bp between each element set, both PA and simulated, and the elements of interest (e.g. human accelerated regions) and an empirical p-value are computed.](Fig1_HAREWorkflow.png "HARE Workflow")
 
 ## Annotation
[…]
