### 2024-04-06 #6359 @ossmith
Dear @nfb1993, @abartlett004, and @jromanowska,

We wanted to thank you again for reviewing our manuscript and software. Below we address the reviewers’ points and discuss the changes we have made in response. Please let us know if you have further thoughts or concerns.

---
_Example of usage: Although the commands are clearly stated and the authors provide some testing scripts to verify installation, there is not a friendly example of how to use the software. I highlight that I was able to run a test example made by myself (I used a T2D publicly available GWAS summary statistics). It would be great if the author could add a small example with some examples of the expected main outputs._

_Example of Usage: I also recommend the inclusion of a small test dataset of inputs and expected outputs for each of the functions. Perhaps the data from Figure 2 from the paper would be suitable._

We have now added an example folder with a bash script for users as well as expected inputs and outputs.

---
_Community guidelines: the author provides where to address potential bugs and issues, however there is not a statement for Software contribution i.e. if somebody external can contribute on the code._

_Community Guidelines: agree with Nicolas as stated above, need guidance for software contribution._

We have revised our Contribution Guidelines to clarify the process for external individuals to add or make changes to the code.

---
_Correction necessary on the repository: The necessary dependencies for installation are listed, but the link to the version of the human genome cache they recommend (homo_sapiens_vep_105_GRCh37.tar.gz) is wrong._

_Installation: as Nicolas mentioned, the link to the recommended human genome cache is wrong. I think it is possible that the intended link is something like https://ftp.ensembl.org/pub/release-105/variation/indexed_vep_cache/homo_sapiens_vep_105_GRCh37.tar.gz, but please confirm._

We have corrected this link.

---
_Correction needed on the software: When trying to run the preranking tool I encountered an error and crashed. The issue was that the time package was not imported. I had to add the line of code “import time” manually to make it work._

We have fixed this bug.

---
_Minor correction in article text: In line 81, “femur rations Using HARE, we decided…”, a dot seems missing._

Resolved.

---
_Side note, not necessary to approve the software: I would not mention Mac OS compatibility, unless it is thoroughly tested. I tried to test the software in an M2 silicon Mac with Ventura OS, I just could not make it run. I was not able to make the conda environment or install ENSEMBL tools. Furthermore, the ENSEMBL tutorial for Mac is outdated and not usable for silicon macs (probably it is for Intel based Macs). I do not think it is the software’s fault. M2 Macs seem to have problems running software in perl (almost not supported there) and certain c++ tools. These problems seem not to be in Intel macs, hence, if possible I would try to test the software in a M1-2 Mac to be sure they are compatible. I was able to run the software very smoothly in a Unix environment._

Thank you for raising this issue! We have revised the README to clarify that, while the software is able to run on Macs where dependencies can be installed, there may be VEP installation issues and the conda environment.yml file may not install it properly. As suggested, we have confirmed that HARE does run on a Mac M2 (Intel) running MacOS Big Sur 11.7.10 with successful installation of VEP.

–-
_@kdm9: And I can give a single overarching review comment for free 😄: Seems like a rather impressive tool, but could the authors also please confirm that the tool works with non-human data. I see no inherent reason it can't, but the entire documentation assumes human data. Perhaps a quick test with e.g. AraGWAS or some other plant dataset would be wise, just to check there are no bugs with non-human data._

We also wanted to address this request from the pre-review thread from @ kdm9. We have added this functionality and tested it on GWAS data from the following non-human species: Bos taurus (vertebrate), Arabadopsis thaliana (plant), Drosophila melanogaster and Apis mellifera (metazoa), and Saccharomyces cerevisiae (fungi). BioMart, a HARE dependency, does not currently support bacteria or protists.

–-
_Functionality: When a test dataset is included, I will be able to evaluate functionality._

Please let us know if you have additional concerns or comments once you are able to evaluate the tool.
