### 2024-05-02 #6359 @abartlett004
With the inclusion of the test data and examples, I am happy to recommend this for publication. Congratulations!

One final note, however--it seems that the files described in `exampleREADME.md` are not exactly what are produced when `example_run.sh` is executed. Specifically, `exampleREADME.md` leads me to expect files with the nomenclature `example_result.*`, whereas what is actually produced are `test_example.*` files. A minor thing, but easy to fix to make it consistent.

Additionally,  for whatever it's worth, I had similar difficulties as Nicolas in terms of installing HARE's dependencies on an M3 Mac running Sonoma, but it was fairly smooth on Windows/WSL. Definitely not the fault of HARE--Apple silicon just doesn't play nicely with the rest of the world, it seems :)
