### 2024-05-06 #6359 @ossmith
Thank you @abartlett004, @nfb1993, and @jromanowska! We appreciate your time and reviews.

> It seems that the files described in exampleREADME.md are not exactly what are produced when example_run.sh is executed. Specifically, exampleREADME.md leads me to expect files with the nomenclature example_result.*, whereas what is actually produced are test_example.* files. A minor thing, but easy to fix to make it consistent.

Thank you for catching this! We've edited the README to clarify that the file stem users should expect - specifically when running the example bash script - is `test_example.*` instead of `example_result.*`, but this is because we don't want users to overwrite the provided output files (example_results.*) by using the same file stem.
