### 2024-05-08 8d43b21
MINOR: Update version for JOSS release
- Update version to 1.2.0 for JOSS release
- Hide paper folder from main branch
