+++ paper/paper.md
@@ -1,67 +1,0 @@
----
-title: 'HARE: A Python workflow for analyzing genomic feature enrichment in GWAS datasets'
-tags:
-  - Python
-  - bioinformatics
-  - evolution
-  - gwas
-  - enrichment analysis
-authors:
-  - name: Olivia S. Smith
-    orcid: 0000-0001-5435-2982
-    affiliation: "1"
-  - name: Eucharist Kun
-    orcid: 0000-0001-9919-173X
-    affiliation: "1"
-  - name: Vagheesh M. Narasimhan
-    orcid: 0000-0001-8651-8844
-    affiliation: "1, 2, 3"
-    corresponding: true
-affiliations:
-  - name: Department of Integrative Biology, The University of Texas at Austin, USA
-    index: 1
-  - name: Department of Statistics and Data Science, The University of Texas at Austin, USA
-    index: 2
-  - name: Department of Population Health, Dell Medical School
-    index: 3
-date: 2 January 2024
-bibliography: paper.bib
----
-
-# Summary
-Genomic datasets have grown by orders of magnitude in the last decade and thus require increasingly flexible and streamlined analysis workflows. One major line of research in genomics is gene enrichment analysis, which aims to quantitatively assess whether the genomic basis of a phenotype of interest has higher association with other genomic regions than would be expected by chance. HARE is a Python pipeline which annotates genome-wide significant positions, identifies overlap between these positions and a provided list of genomic features (here called 'elements') of interest, and determines enrichment likelihood through a resampling process of random length-matched regions. HARE is written in Python and is available for installation from source or via the Python Package Index (pip).
-
-# Statement of need
-HARE is a computational pipeline run via command line for analyzing enrichment of element sets, such as the set of human-accelerated regions (HARs), in genomic regions associated with a phenotype of interest. Several genetic enrichment approaches are already available in the literature, such as gene set enrichment analysis (GSEA) [@Subramanian2005; @Mootha2003], stratified linkage disequilibrium score regression (LDSC) [@Finucane2015], and FUMA GWAS [@Wantanabe2017]. However, these approaches are not translatable to studies examining overlaps in custom regions of the genome or when these regions are short, which is known to lead to bias in some approaches [@Finucane2015]. Our pipeline is flexible on either end of the enrichment analysis, both in terms of the regions it utilizes as well as the type of input genomic scan provided, allowing for a completely flexible model for genomic enrichment analysis.
-
-Many tools also require time-consuming data reformatting which HARE minimizes by accepting a variety of summary statistic file formats. HARE dependencies can be installed with conda and pip, allowing for an immediate run of the complete analysis pipeline. HARE has already been used to analyze over 200 traits in published work as well as ongoing studies presented at conferences [@Kun2023a; @Kun2023b; @Xu2023], exemplifying its usefulness in current and future research.
-
-# Dependencies
-HARE requires Ensembl's Variant Effect Predictor [@McLaren2016] with the appropriate cache ([VEP Cache Documentation](https://uswest.ensembl.org/info/docs/tools/vep/script/vep_cache.html#cache)) as well as BEDTools [@Quinlan2010]. It also requires the use of `wget` for querying Ensembl BioMart [@Martin2023]. It is recommended that users install dependencies through a combination of conda using the `environment.yml` file included in the repository and pip. Complete instructions for dependency and tool installation can be found in the repository README.
-
-# Workflow
-A visual depiction of the HARE workflow is shown in Figure 1.
-
[…]
