+++ paper/paper.bib
@@ -1,230 +1,0 @@
-@ARTICLE{Subramanian2005,
-  title = {"Gene set enrichment analysis: A knowledge-based approach for interpreting genome-wide expression profiles"},
-  author = {Aravind Subramanian and Pablo Tamayo and Vamsi K. Mootha and Sayan Mukherjee and Benjamin L. Ebert and Michael A. Gillette and Amanda Paulovich and Scott L. Pomeroy and Todd R. Golub and Eric S. Lander and Jill P. Mesirov},
-  journal = {PNAS},
-  volume = {102},
-  issue = {43},
-  pages = {15545-15550},
-  year = {2005},
-  doi = {10.1073/pnas.0506580102},
-  url = {https://doi.org/10.1073/pnas.0506580102}
-  }
-
-@ARTICLE{Mootha2003,
-  title = {"PGC-1α-responsive genes involved in oxidative phosphorylation are coordinately downregulated in human diabetes."},
-  author = {Vamsi K. Mootha and Cecilia M. Lindgren and Karl-Fredrik Eriksson and Aravind Subramanian and Smita Sihag and Joseph Lehar and Pere Puigserver and Emma Carlsson and Martin Ridderstråle and Esa Laurila and Nicholas Houstis and Mark J. Daly and Nick Patterson and Jill P. Mesirov and Todd R. Golub and Pablo Tamayo and Bruce Spiegelman and Eric S. Lander and Joel N. Hirschhorn and David Altshuler and Leif C. Groop},
-  journal = {Nat Genet},
-  volume = {34},
-  pages = {267-273},
-  year = {2003},
-  doi = {10.1038/ng1180},
-  url = {https://doi.org/10.1038/ng1180}
-  }
-
-@ARTICLE{Finucane2015,
-  title = {"Partitioning heritability by functional annotation using genome-wide association summary statistics"},
-  author = {Hilary K. Finucane and Brendan Bulik-Sullivan and Alexander Gusev and Gosia Trynka and Yakir Reshef and Po-Ru Loh and Verneri Anttila and Han Xu and Chongzhi Zang and Kyle Farh and Stephan Ripke and Felix R. Day and ReproGen Consortium and Schizophrenia Working Group of the Psychiatric Genomics Consortium and The RACI Consortium and Shaun Purcell and Eli Stahl and Sara Lindstrom and John R.B. Perry and Yukinori Okada and Soumya Raychaudhuri and Mark J. Daly and Nick Patterson and Benjamin M. Neale and Alkes L. Price},
-  journal = {Nat Genet},
-  volume = {47},
-  pages = {1228-1235},
-  year = {2015},
-  doi = {10.1038/ng.3404},
-  url = {https://doi.org/10.1038/ng.3404}
-  }
-
-@ARTICLE{Wantanabe2017,
-  title = {"Functional mapping and annotation of genetic associations with FUMA"},
-  author = {Kyoko Watanabe and Erdogan Taskesen and Arjen van Bochoven and Danielle Posthuma},
-  journal = {Nat Commun},
-  volume = {8},
-  pages = {1826},
-  year = {2017},
-  doi = {10.1038/s41467-017-01261-5},
-  url = {https://doi.org/10.1038/s41467-017-01261-5}
-}
-
-@ARTICLE{Grüning2018,
-  title = {"Bioconda: sustainable and comprehensive software distribution for the life sciences"},
[…]
