### 2024-05-08 #6359 @ossmith
Hi @jromanowska, we just wanted to follow up that we've completed the following tasks:

**The version number for release is v1.2.0**
**The DOI is 10.5281/zenodo.11154700**

Please let us know if you need anything else.
