### 2024-05-13 f8b6151
MINOR: Update paper.md
