### 2024-05-13 #6359 @ossmith
Hi @jromanowska - ah ok! We've removed the 'corresponding author' designation. We've also reviewed the guidelines and the other authors do meet the conditions for non-code-related authorship (e.g. project direction and code testing). Perhaps you already noticed this as well but the missing DOIs above are for conference proceeding citations and not articles.
